se agregan msgs genericos de respuesta

+++ database/migrations/2022_07_24_212324_create_tbl_informes_actividades_table.php
@@ -15,7 +15,7 @@ class CreateTblInformesActividadesTable extends Migration
     {
         Schema::create('tbl_informes_actividades', function (Blueprint $table) {
             $table->bigIncrements('id_informe_actividad');
-            $table->integer('id_actividad')->nullable();
+            $table->unsignedBigInteger('id_actividad')->nullable();
             $table->unsignedBigInteger('id_usuareg');
             $table->string('link');
             $table->timestamps();
